do load.sh at runtime rather than build time

+++ virtuosodata-frombackup/load.sh
@@ -1,6 +1,15 @@
 #!/bin/sh
+set -e
 cd /virtuoso
+if [ -f .loaded ] ; then
+  exit
+fi
+if ! [ -f *#1.bp ] ; then
+  /usr/local/bin/download.sh	
+fi
 DUMP=`echo *#1.bp | sed s/1.bp//`
 echo Loading $DUMP
 virtuoso-t +configfile /etc/virtuoso-opensource-7/virtuoso.ini  +restore-backup $DUMP
 rm $DUMP*
+touch .loaded
+echo "Loading completed"
